商品削除機能(修正済)
自分の出品した商品以外の商品を削除しようとすると、詳細ページにリダイレクトするように修正しました。

+++ app/controllers/products_controller.rb
@@ -39,6 +39,9 @@ class ProductsController < ApplicationController
 
   def destroy
     product = Product.find(params[:id])
+    unless product.user == current_user
+      redirect_to product_path
+    end
     product.destroy
     redirect_to root_path
